refactor(bundle): extract config tree and parameter registration for reuse
Move the configuration tree definition out of
SymfonySecurityAuditorBundle::configure() into a dedicated
AuditConfigurationDefinition, and the ~45 container-parameter
assignments out of loadExtension() into a ContainerParameterRegistrar.
Add an AuditConfiguration (ConfigurationInterface) that wraps the shared
tree so a standalone Symfony Processor can normalize the same schema the
bundle uses.

This is internal groundwork toward distributing the auditor as a
standalone tool (#76): both the bundle and the future standalone
bootstrap now build their configuration from one shared source, so the
normalized output cannot drift between them. No behavior change — the
bundle produces the same container parameters as before, guarded by the
existing SymfonySecurityAuditorBundleTest.
