+++ src/Audit/Infrastructure/Config/AuditConfiguration.php
@@ -1,0 +1,31 @@
+<?php
+
+/*
+ * This file is part of the vinceamstoutz/symfony-security-auditor package.
+ *
+ * (c) Vincent Amstoutz <vincent.amstoutz.dev@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace VinceAmstoutz\SymfonySecurityAuditor\Audit\Infrastructure\Config;
+
+use Symfony\Component\Config\Definition\Builder\TreeBuilder;
+use Symfony\Component\Config\Definition\ConfigurationInterface;
+
+final readonly class AuditConfiguration implements ConfigurationInterface
+{
+    public const string ROOT_NODE = 'symfony_security_auditor';
+
+    public function getConfigTreeBuilder(): TreeBuilder
+    {
+        $treeBuilder = new TreeBuilder(self::ROOT_NODE);
+
+        (new AuditConfigurationDefinition())->defineChildren($treeBuilder->getRootNode()->children());
+
+        return $treeBuilder;
+    }
+}
