+++ src/Audit/Infrastructure/Config/AuditConfigurationDefinition.php
@@ -1,0 +1,360 @@
+<?php
+
+/*
+ * This file is part of the vinceamstoutz/symfony-security-auditor package.
+ *
+ * (c) Vincent Amstoutz <vincent.amstoutz.dev@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace VinceAmstoutz\SymfonySecurityAuditor\Audit\Infrastructure\Config;
+
+use Symfony\Component\Config\Definition\Builder\NodeBuilder;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Application\Agent\AttackerAgent;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Application\Agent\AuditOrchestrator;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Application\Agent\ReviewerAgent;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Domain\Model\VulnerabilityType;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Infrastructure\FileSystem\ProjectFileScanner;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Infrastructure\LLM\RetryPolicy;
+
+/**
+ * @internal
+ */
+final readonly class AuditConfigurationDefinition
+{
+    public function defineChildren(NodeBuilder $nodeBuilder): void
+    {
+        $nodeBuilder
+                ->scalarNode('model')
+                    ->defaultValue('claude-opus-4-8')
+                    ->info('Model name for both Attacker and Reviewer. Must be supported by the configured platform.')
+                ->end()
+                ->enumNode('profile')
+                    ->values(['fast', 'balanced', 'thorough'])
+                    ->defaultValue('balanced')
+                    ->info('One-knob preset bundling the cost/speed/depth levers; any explicitly configured key always wins over the profile. `fast`: one attacker iteration, lean pre-scan (marker-bearing files only), code slicing on, up to 4 concurrent reviewer calls. `balanced` (default): identical to configuring nothing. `thorough`: balanced plus PoC synthesis for high-severity validated findings.')
+                ->end()
+                ->scalarNode('attacker_model')
+                    ->defaultNull()
+                    ->info('Override: dedicated model for the Attacker role. Falls back to `model` when null.')
+                ->end()
+                ->scalarNode('reviewer_model')
+                    ->defaultNull()
+                    ->info('Override: dedicated model for the Reviewer role. Falls back to `model` when null.')
+                ->end()
+                ->integerNode('max_output_tokens')
+                    ->defaultValue(4096)
+                    ->min(1)
+                    ->info("Maximum output tokens per LLM call for both Attacker and Reviewer. Sets `max_tokens` in every platform request. Default 4096; symfony/ai's Anthropic bridge otherwise defaults to a much smaller value (~1000) that silently truncates findings.")
+                ->end()
+                ->integerNode('attacker_max_output_tokens')
+                    ->defaultNull()
+                    ->min(1)
+                    ->info('Override: dedicated max output tokens for the Attacker role. Falls back to `max_output_tokens` when null. Useful when the attacker needs more headroom for detailed `record_vulnerability` tool-call arguments.')
+                ->end()
+                ->integerNode('reviewer_max_output_tokens')
+                    ->defaultNull()
+                    ->min(1)
+                    ->info('Override: dedicated max output tokens for the Reviewer role. Falls back to `max_output_tokens` when null.')
+                ->end()
+                ->booleanNode('provider_json_mode')
+                    ->defaultFalse()
+                    ->info('Opt into the provider-native JSON mode by sending `response_format: {type: json_object}` on every LLM call. Honored by OpenAI/Mistral/Ollama; silently ignored by Anthropic (which has no equivalent knob). Default false because behaviour is provider-dependent — only enable if your provider supports it. The prompt contract ("Return ONLY the JSON array") remains authoritative.')
+                ->end()
+                ->arrayNode('scan')
+                    ->addDefaultsIfNotSet()
+                    ->children()
+                        ->arrayNode('included_paths')
+                            ->info('Project-relative directories and files that define the scan surface — the sole scoping knob. Defaults to the Symfony Flex skeleton (src/, config/, templates/, public/index.php). Anything outside this list is silently skipped, including vendor/, node_modules/, var/, tests/, migrations/, build artefacts, IDE folders, and any other top-level tree. Tighten or extend the list to match non-standard layouts; the audit only inspects what is listed here.')
+                            ->scalarPrototype()->end()
+                            ->defaultValue(ProjectFileScanner::DEFAULT_INCLUDED_PATHS)
+                        ->end()
+                        ->booleanNode('respect_gitignore')
+                            ->defaultTrue()
+                            ->info('When true, files ignored by the project .gitignore are excluded from the scan. Default true — matches the host project intent (committed code only) and avoids analyzing generated/cached artefacts. Set false for full-tree scans (rare).')
+                        ->end()
+                        ->integerNode('max_file_size_kb')
+                            ->defaultValue(ProjectFileScanner::DEFAULT_MAX_FILE_SIZE_KB)
+                            ->min(1)
+                            ->info('Skip files larger than this size, in kilobytes.')
+                        ->end()
+                        ->arrayNode('custom_risk_patterns')
+                            ->info('Project-specific risk markers merged into the deterministic pre-scanner. Keyed by file-type bucket (controller, voter, entity, repository, form, template, config, php, authenticator, messenger_handler, webhook_consumer, event_subscriber, normalizer, scheduler). Each entry is `<label>: { regex: <PCRE>, description: <human text> }`. Surface team idioms the built-in patterns do not know about (e.g. "must call AuditService::log() after every privileged action").')
+                            ->useAttributeAsKey('bucket')
+                            ->arrayPrototype()
+                                ->useAttributeAsKey('label')
+                                ->arrayPrototype()
+                                    ->children()
+                                        ->scalarNode('regex')->isRequired()->cannotBeEmpty()->info('PCRE pattern with delimiters, matched per-line against file content.')->end()
+                                        ->scalarNode('description')->isRequired()->cannotBeEmpty()->info('Short human description rendered in the prompt next to the marker.')->end()
+                                    ->end()
+                                ->end()
+                            ->end()
+                            ->defaultValue([])
+                        ->end()
+                        ->arrayNode('secret_scrubbing')
+                            ->addDefaultsIfNotSet()
+                            ->info('Redact credential-shaped strings from file content before it reaches the LLM. Covers AWS/GitHub/Stripe/Slack/Google API keys, JWTs, PEM private keys, and env-style credential assignments.')
+                            ->children()
+                                ->booleanNode('enabled')
+                                    ->defaultTrue()
+                                    ->info('When true, file content is run through the configured scrubber before chunking. Default true — credentials in committed sample configs or .env.dist files would otherwise be sent verbatim to the LLM provider.')
+                                ->end()
+                                ->arrayNode('additional_patterns')
+                                    ->info('Extra PCRE patterns merged with the defaults. Use to redact project-specific tokens (e.g. internal API keys).')
+                                    ->scalarPrototype()->end()
+                                    ->defaultValue([])
+                                ->end()
+                            ->end()
+                        ->end()
+                    ->end()
+                ->end()
+                ->arrayNode('audit')
+                    ->addDefaultsIfNotSet()
+                    ->children()
+                        ->integerNode('max_iterations')
+                            ->defaultNull()
+                            ->min(1)
+                            ->info('Maximum number of attacker/reviewer iterations per audit. Defaults to the active profile (balanced/thorough: 3, fast: 1).')
+                        ->end()
+                        ->floatNode('min_confidence')
+                            ->defaultValue(AuditOrchestrator::DEFAULT_MIN_CONFIDENCE)
+                            ->min(0.0)
+                            ->max(1.0)
+                            ->info('Minimum attacker self-reported confidence (0.0–1.0) required to forward a finding to the reviewer.')
+                        ->end()
+                        ->integerNode('reviewer_batch_size')
+                            ->defaultValue(ReviewerAgent::DEFAULT_BATCH_SIZE)
+                            ->min(1)
+                            ->info('Number of findings reviewed per LLM call. 1 = one finding per call (highest precision, highest latency). Higher values reduce cost and latency at the risk of cross-talk between findings in the prompt.')
+                        ->end()
+                        ->booleanNode('tools_enabled')
+                            ->defaultTrue()
+                            ->info('Give the attacker access to tools (read_file, grep, list_files, lookup_advisory) for cross-file investigation. Default true — without tools, lookup_advisory is dead weight and the attacker is blind across files. Costs more LLM round-trips per chunk; combine with Anthropic prompt caching (`cache_retention` in `ai.yaml`) to offset the input-token cost.')
+                        ->end()
+                        ->booleanNode('structured_collection')
+                            ->defaultTrue()
+                            ->info("When true (default), the attacker emits findings by calling a schema-enforced `record_vulnerability` tool, one call per finding, instead of returning a JSON array. The platform validates each call against the tool's input schema, so malformed shapes (bare strings like \"dev\"/\"test\", wrapper objects like `{\"vulnerabilities\": [...]}` ) become structurally impossible. Works across every provider that supports tool use (Anthropic, OpenAI, Mistral, Ollama with tool-capable models). Set to false to fall back to the tightened JSON-array prompt path.")
+                        ->end()
+                        ->booleanNode('reviewer_structured_collection')
+                            ->defaultTrue()
+                            ->info('When true (the default), the reviewer records each verdict by calling a schema-enforced `record_review` tool instead of returning a JSON array, so a malformed verdict never costs a discarded (but fully billed) response. Verdicts are served from and stored to the reviewer-verdict cache exactly like the JSON path. The explicit opt-ins `reviewer_tools_enabled: true` and `reviewer_max_concurrent` > 1 take precedence and keep the JSON path. Set false to force JSON-array output (for models without tool-use support).')
+                        ->end()
+                        ->booleanNode('stable_system_prompt')
+                            ->defaultTrue()
+                            ->info('When true (the default), the attacker emits the full expert skill set in its system prompt for every chunk, instead of only the skills matching the chunk\'s file types. This makes the system-prompt prefix byte-identical across chunks so provider prompt caching reads it on every call after the first — Anthropic (`cache_retention` in `ai.yaml`, default `short`), OpenAI, Gemini, and DeepSeek all cache prompt prefixes; a large input-token saving on multi-chunk audits. Set false (relevance-only skills, smaller prompt) for providers without prompt caching.')
+                        ->end()
+                        ->integerNode('max_tool_iterations')
+                            ->defaultValue(AttackerAgent::DEFAULT_MAX_TOOL_ITERATIONS)
+                            ->min(1)
+                            ->info('Maximum tool-call rounds per chunk before forcing the attacker to commit to a final answer. Bounds runaway tool use.')
+                        ->end()
+                        ->booleanNode('reviewer_tools_enabled')
+                            ->defaultFalse()
+                            ->info('Give the reviewer access to the same tool registry the attacker uses, so it can verify cross-file context (parent-class guards, access_control rules, upstream sanitizers) instead of guessing from the Full File Context alone. Default false — adds round-trips per finding; opt-in for high-precision audits.')
+                        ->end()
+                        ->integerNode('attacker_max_concurrent')
+                            ->defaultNull()
+                            ->min(1)
+                            ->info('Maximum attacker chunk analyses resolved concurrently in the default structured-collection mode (audit.structured_collection: true), when the configured platform exposes an async transport. The attacker phase is usually the longest; setting this to 4-8 (within your provider rate limit) cuts it proportionally. Defaults to the active profile (balanced/thorough: 1 — sequential, fast: 4). Ignored when audit.tools_enabled gives the attacker a cross-file tool registry or structured_collection is off.')
+                        ->end()
+                        ->integerNode('reviewer_max_concurrent')
+                            ->defaultNull()
+                            ->min(1)
+                            ->info('Maximum reviewer LLM calls resolved concurrently when reviewing one finding per call (reviewer_batch_size <= 1) with reviewer tools off. The reviewer phase is often half the audit wall-clock; setting this to 4-8 (within your provider rate limit) cuts it proportionally. Defaults to the active profile (balanced/thorough: 1 — sequential, fast: 4). Ignored when reviewer tools are enabled or the configured platform has no async transport.')
+                        ->end()
+                        ->integerNode('reviewer_max_tool_iterations')
+                            ->defaultValue(ReviewerAgent::DEFAULT_MAX_TOOL_ITERATIONS)
+                            ->min(1)
+                            ->info("Maximum tool-call rounds per finding before forcing the reviewer to commit to a verdict. Lower default than the attacker because the reviewer's job is verification, not exploration.")
+                        ->end()
+                        ->scalarNode('baseline')
+                            ->defaultNull()
+                            ->info('Default path to a baseline file of accepted-finding fingerprints. Findings whose fingerprint is listed are suppressed from the report and excluded from the exit code, so previously-accepted findings no longer fail CI. The --baseline CLI option overrides this path; --generate-baseline writes the file. Null (default) disables baselining.')
+                        ->end()
+                        ->enumNode('fail_on')
+                            ->values(['safe', 'low', 'medium', 'high', 'critical'])
+                            ->defaultValue('critical')
+                            ->info("Minimum aggregate risk level that makes `audit:run` exit 1 (the CI gate). The audit exits 1 when the report's risk level is at or above this threshold, 0 otherwise (a budget abort still exits 2). Default `critical` preserves the historical behaviour (only a CRITICAL risk level fails). Set `high`/`medium`/`low` to fail PRs earlier. `safe` fails on every completed audit. The --fail-on CLI option overrides this per run.")
+                        ->end()
+                        ->arrayNode('excluded_types')
+                            ->info('Vulnerability types dropped from the report and the exit code, even when validated. Mutes a noisy class (e.g. missing_rate_limiting) without enumerating per-finding baseline fingerprints. Each value must be a VulnerabilityType (sql_injection, missing_voter, …). Empty (default) mutes nothing. Wins over included_types.')
+                            ->enumPrototype()->values($this->vulnerabilityTypeValues())->end()
+                            ->defaultValue([])
+                        ->end()
+                        ->arrayNode('included_types')
+                            ->info('Allowlist of vulnerability types: when non-empty, only findings whose type is listed are reported and counted toward the exit code. Each value must be a VulnerabilityType. Empty (default) includes every type. excluded_types still wins over this list.')
+                            ->enumPrototype()->values($this->vulnerabilityTypeValues())->end()
+                            ->defaultValue([])
+                        ->end()
+                        ->arrayNode('static_prescan')
+                            ->addDefaultsIfNotSet()
+                            ->info('Deterministic zero-token risk-marker scan that runs before the LLM. Flags concrete locations (unserialize, |raw, csrf_protection: false, hardcoded secrets, Doctrine string concatenation, etc.) so the attacker prompt can focus on them. In lean mode, files with zero markers are skipped entirely — biggest token saver on large codebases.')
+                            ->children()
+                                ->booleanNode('enabled')
+                                    ->defaultTrue()
+                                    ->info('When true, every chunk is preceded by a "Pre-Scan Risk Markers" section in the user message. Default true — pure win on detection quality, zero token cost for the scan itself.')
+                                ->end()
+                                ->booleanNode('lean_mode')
+                                    ->defaultNull()
+                                    ->info("When true, files with zero markers are dropped before the LLM ever sees them. Slashes token spend on real codebases (often 40-70%) at the cost of missing patterns the regex pre-scanner doesn't know about. Defaults to the active profile (balanced/thorough: false, fast: true).")
+                                ->end()
+                            ->end()
+                        ->end()
+                        ->arrayNode('chunking')
+                            ->addDefaultsIfNotSet()
+                            ->info('How files are grouped into LLM calls. `feature` (default) packs a controller with its entity/repository/form/voter/templates so the LLM can follow cross-file data flow — the biggest detection-quality win. `type` keeps the legacy behaviour of sorting by attack-surface priority and slicing into fixed-size windows.')
+                            ->children()
+                                ->enumNode('strategy')
+                                    ->values(['feature', 'type'])
+                                    ->defaultValue('feature')
+                                    ->info('Chunking strategy. `feature` colocates related files (UserController + User entity + UserRepository + …) in one chunk; `type` chunks by file-type priority. Default `feature`.')
+                                ->end()
+                            ->end()
+                        ->end()
+                        ->arrayNode('escalation')
+                            ->addDefaultsIfNotSet()
+                            ->info('Two-pass attacker. A cheap-model sweep runs on every chunk; the expensive model only re-analyses files the cheap sweep flagged. Cuts attacker token spend ~3-5x on real projects where most files are inert, with detection quality close to running the expensive model on everything. Off by default — opt-in for cost-sensitive audits.')
+                            ->children()
+                                ->booleanNode('enabled')
+                                    ->defaultFalse()
+                                    ->info('When true, AttackerAgentInterface is wired to the EscalatingAttackerAgent wrapper. Requires `cheap_model` to be set.')
+                                ->end()
+                                ->scalarNode('cheap_model')
+                                    ->defaultNull()
+                                    ->info('Provider model id used for the cheap first pass (e.g. claude-haiku-4-5-20251001, gpt-5-mini, mistral-small). Falls back to the reviewer model when null.')
+                                ->end()
+                            ->end()
+                        ->end()
+                        ->arrayNode('code_slicing')
+                            ->addDefaultsIfNotSet()
+                            ->info('Trim large PHP files down to security-relevant slices before they reach the LLM. The slicer keeps imports, attributes, class signatures, properties, and the FULL body of methods that touch security-relevant tokens (Request, Doctrine query builder, unserialize, shell exec, mailer, HttpClient, …). All other lines are replaced one-for-one with a `// elided` placeholder so line numbers stay accurate. Typical saving: 50-70% input tokens on controllers / services over 100 lines.')
+                            ->children()
+                                ->booleanNode('enabled')
+                                    ->defaultNull()
+                                    ->info('When true, the configured CodeSlicerInterface runs over every chunked file. Default false — opt-in for cost-sensitive audits; smaller files and unfamiliar idioms keep more signal when sent unsliced.')
+                                ->end()
+                                ->integerNode('min_lines_before_slicing')
+                                    ->defaultValue(80)
+                                    ->min(10)
+                                    ->info('Skip slicing for files shorter than this. Below ~80 lines the saving is not worth the missing context.')
+                                ->end()
+                            ->end()
+                        ->end()
+                        ->arrayNode('poc_synthesis')
+                            ->addDefaultsIfNotSet()
+                            ->info('Optional follow-up stage that generates a concrete, copy-pasteable proof-of-concept (curl command, console invocation, payload body) for every validated finding at or above the configured severity floor. Spends extra LLM tokens per finding; off by default. Turn on when shipping reports to engineers who need actionable reproduction steps.')
+                            ->children()
+                                ->booleanNode('enabled')
+                                    ->defaultNull()
+                                    ->info('When true, the PoCSynthesisStage runs after the audit and attaches a `synthesized_poc` field to qualifying findings. Default false.')
+                                ->end()
+                                ->enumNode('severity_floor')
+                                    ->values(['critical', 'high', 'medium', 'low', 'info'])
+                                    ->defaultValue('high')
+                                    ->info('Minimum severity that triggers PoC synthesis. Default `high` — synthesize for critical+high only; medium and below keep the attacker\'s original proof string.')
+                                ->end()
+                            ->end()
+                        ->end()
+                        ->arrayNode('budget')
+                            ->addDefaultsIfNotSet()
+                            ->info('Hard ceiling on cumulative LLM usage per audit run. Aborts the audit cleanly (exit code 2) with the partial report instead of running away on cost.')
+                            ->children()
+                                ->integerNode('max_tokens')
+                                    ->defaultNull()
+                                    ->min(1)
+                                    ->info('Maximum total tokens (input + output, across attacker + reviewer) before the audit aborts. `null` (default) = unlimited.')
+                                ->end()
+                                ->floatNode('max_cost_usd')
+                                    ->defaultNull()
+                                    ->info('Maximum estimated cost (USD) before the audit aborts. Computed via the configured `PricingProviderInterface`. `null` (default) = unlimited.')
+                                ->end()
+                            ->end()
+                        ->end()
+                        ->arrayNode('retry')
+                            ->addDefaultsIfNotSet()
+                            ->info('Bounded exponential-backoff retry around every LLM call. Transient failures (provider 429/5xx, network blips) are retried with jittered delays; non-transient failures (auth, validation) fail fast.')
+                            ->children()
+                                ->integerNode('max_attempts')
+                                    ->defaultValue(RetryPolicy::DEFAULT_MAX_ATTEMPTS)
+                                    ->min(1)
+                                    ->info('Total attempts per LLM call, including the first try. `1` disables retries.')
+                                ->end()
+                                ->integerNode('initial_delay_ms')
+                                    ->defaultValue(RetryPolicy::DEFAULT_INITIAL_DELAY_MS)
+                                    ->min(0)
+                                    ->info('Base delay (milliseconds) before the first retry. Subsequent retries multiply by `backoff_multiplier`.')
+                                ->end()
+                                ->floatNode('backoff_multiplier')
+                                    ->defaultValue(RetryPolicy::DEFAULT_BACKOFF_MULTIPLIER)
+                                    ->min(1.0)
+                                    ->info('Exponential growth factor between retries. With initial 500ms and multiplier 2.0, retries wait ~500, ~1000, ~2000 ms.')
+                                ->end()
+                                ->floatNode('jitter_ratio')
+                                    ->defaultValue(RetryPolicy::DEFAULT_JITTER_RATIO)
+                                    ->min(0.0)
+                                    ->max(1.0)
+                                    ->info('Jitter applied to each computed delay, as a fraction in `[0.0, 1.0]`. `0.2` means each delay varies within ±20% of the base.')
+                                ->end()
+                            ->end()
+                        ->end()
+                        ->arrayNode('rate_limit')
+                            ->addDefaultsIfNotSet()
+                            ->info('Proactive token-bucket throttle around every LLM call. Each dimension is independently nullable; when all three are `null` (the default) the bundle wires `NullRateLimiter` and behavior matches the pre-existing retry-only path. Configure the limits enforced by your provider tier (e.g. Anthropic RPM/ITPM/OTPM) to avoid hitting 429 in the first place — exhausted retries will still surface, but the steady-state path stays inside quota.')
+                            ->children()
+                                ->integerNode('requests_per_minute')
+                                    ->defaultNull()
+                                    ->min(1)
+                                    ->info('Maximum LLM requests per minute. `null` (default) disables this dimension.')
+                                ->end()
+                                ->integerNode('input_tokens_per_minute')
+                                    ->defaultNull()
+                                    ->min(1)
+                                    ->info('Maximum input tokens per minute. `null` (default) disables this dimension. A single request whose estimated input exceeds this cap throws `RateLimitRequestTooLargeException`.')
+                                ->end()
+                                ->integerNode('output_tokens_per_minute')
+                                    ->defaultNull()
+                                    ->min(1)
+                                    ->info('Maximum output tokens per minute. `null` (default) disables this dimension. Counted post-hoc from `record()` so the next `acquire()` defers until the window resets when the bucket is full.')
+                                ->end()
+                            ->end()
+                        ->end()
+                    ->end()
+                ->end()
+                ->arrayNode('cache')
+                    ->addDefaultsIfNotSet()
+                    ->children()
+                        ->booleanNode('enabled')
+                            ->defaultTrue()
+                            ->info('Enable content-hash cache for attacker chunks. Skips the LLM call when an identical chunk has been analyzed before. Default true — huge cost saver on repeated runs (CI, PR scans).')
+                        ->end()
+                        ->scalarNode('dir')
+                            ->defaultValue('%kernel.cache_dir%/symfony_security_auditor/attacker')
+                            ->info('Filesystem path for the attacker cache. Created on first write.')
+                        ->end()
+                        ->booleanNode('prompt_caching')
+                            ->defaultTrue()
+                            ->setDeprecated(
+                                'vinceamstoutz/symfony-security-auditor',
+                                '1.7',
+                                'The "%node%" option is deprecated and no longer has any effect. Prompt caching is controlled by your Symfony AI platform: set `cache_retention` (none|short|long) on the anthropic platform in `ai.yaml` (default `short` already enables it); OpenAI and Gemini cache automatically.',
+                            )
+                            ->info('Deprecated and ignored since 1.7. Prompt caching is configured on the Symfony AI platform, not here: set `cache_retention` (none|short|long) on the anthropic platform in `ai.yaml`. OpenAI and Gemini cache automatically.')
+                        ->end()
+                    ->end()
+                ->end()
+        ;
+    }
+
+    /**
+     * @return list<string>
+     */
+    private function vulnerabilityTypeValues(): array
+    {
+        return array_map(static fn (VulnerabilityType $vulnerabilityType): string => $vulnerabilityType->value, VulnerabilityType::cases());
+    }
+}
